+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Note: Can be used with `oxsecurity/megalinter@beta` in your GitHub Action mega-l
   - Config generator tool now supports Drone CI
   - Added information about how to change trigger rules for Drone CI workflow
 - New MegaLinter plugin: [linkcheck](https://github.com/shiranr/linkcheck): Plugin to check and validate Markdown links.
+- Use `--squash` option with docker build in order to reduce images size
 
 - Linter versions upgrades
   - [csharpier](https://csharpier.com/) from 0.21.0 to **0.16.0** on 2022-12-29